+++ metadata.rb
@@ -1,7 +1,7 @@
 name             'chef_phpbrew'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
-license          'All rights reserved'
-description      'Installs/Configures chef_phpbrew'
+maintainer       'c-styles'
+maintainer_email 'eisuke@c-styles.jp'
+license          'MIT License'
+description      'Installs/Configures phpbrew'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'